refactor(packing): split pack detail page
- ♻️ Move pack detail state into a focused composable
- ♻️ Extract pack overview, planning, entries, and delete dialog components
- ♻️ Reuse shared danger action, summary header, and pack identity UI
- ✅ Keep pack, catalog, inventory, and account flows covered by checks
