+++ app/types/packing.ts
@@ -23,6 +23,16 @@ interface PackingListEntry {
   updatedAt: string;
 }
 
+interface PackingListEntryView extends PackingListEntry {
+  detailText: string;
+  displayName: string;
+  isDisabled: boolean;
+  isLoading: boolean;
+  stateClass: 'packed' | 'unpacked';
+  stateText: string;
+  toggleLabel: string;
+}
+
 interface PackingListDetail {
   createdAt: string;
   entries: PackingListEntry[];
@@ -35,10 +45,14 @@ interface PackingListView extends PackingListSummary {
   formattedUpdatedAt: string;
 }
 
+type PackDetailMode = 'planning' | 'checklist'
+
 export type {
+  PackDetailMode,
   PackingListDetail,
   PackingListEntry,
   PackingListEntryInventory,
+  PackingListEntryView,
   PackingListSummary,
   PackingListView
 }
